Fix Vim buffer handling
Each open file must have it's own bim buffer. We don't want to switch a
notebook tab and find that the state of the Vim label has stayed
the same. This change amkes that change, and any other changes
that need to be made as a result of this design change